adds on_duplicate_update (#84)
* adds on_duplicate_update
* corrects updated_at for update on duplicate type
* Refactored existing media check
* Updated onDuplicate Test
* pluck only key on duplicate
* verbose variable names
* Updated Duplicate Docs

+++ src/MediaUploader.php
@@ -25,6 +25,7 @@ class MediaUploader
     const ON_DUPLICATE_REPLACE = 'replace';
     const ON_DUPLICATE_INCREMENT = 'increment';
     const ON_DUPLICATE_ERROR = 'error';
+    const ON_DUPLICATE_UPDATE = 'update';
 
     /**
      * @var FileSystemManager
@@ -248,6 +249,16 @@ class MediaUploader
     }
 
     /**
+     * Overwrite existing file when file already exists at destination.
+     *
+     * @return static
+     */
+    public function onDuplicateUpdate()
+    {
+        return $this->setOnDuplicateBehavior(self::ON_DUPLICATE_UPDATE);
+    }
+
+    /**
      * Append incremented counter to file name when file already exists at destination.
      *
      * @return static
@@ -258,7 +269,7 @@ class MediaUploader
     }
 
     /**
-     * Overwrite existing file when file already exists at destination.
+     * Overwrite existing Media when file already exists at destination.
      *
      * @return static
      */
@@ -443,7 +454,7 @@ class MediaUploader
         $model->directory = $this->directory;
         $model->filename = $this->generateFilename();
 
-        $this->verifyDestination($model);
+        $model = $this->handleDuplicate($model);
 
         if (is_callable($this->before_save)) {
             call_user_func($this->before_save, $model, $this->source);
@@ -647,20 +658,6 @@ class MediaUploader
     }
 
     /**
-     * Verify that the intended destination is available and handle any duplications.
-     * @param  \Plank\Mediable\Media  $model
-     * @return void
-     */
-    private function verifyDestination(Media $model)
-    {
-        $storage = $this->filesystem->disk($model->disk);
-
-        if ($storage->has($model->getDiskPath())) {
-            $this->handleDuplicate($model);
-        }
-    }
-
-    /**
      * Decide what to do about duplicated files.
      * @param  \Plank\Mediable\Media  $model
      * @return void
@@ -668,17 +665,27 @@ class MediaUploader
      */
     private function handleDuplicate(Media $model)
     {
-        switch ($this->config['on_duplicate']) {
-            case static::ON_DUPLICATE_ERROR:
-                throw FileExistsException::fileExists($model->getDiskPath());
-                break;
-            case static::ON_DUPLICATE_REPLACE:
-                $this->deleteExistingMedia($model);
-                break;
-            case static::ON_DUPLICATE_INCREMENT:
-            default:
-                $model->filename = $this->generateUniqueFilename($model);
+        if ($this->filesystem->disk($model->disk)->has($model->getDiskPath())) {
+            switch ($this->config['on_duplicate']) {
+                case static::ON_DUPLICATE_ERROR:
+                    throw FileExistsException::fileExists($model->getDiskPath());
+                    break;
+                case static::ON_DUPLICATE_REPLACE:
+                    $this->deleteExistingMedia($model);
+                    break;
+                case static::ON_DUPLICATE_UPDATE:
+                    $this->deleteExistingFile($model);
+                    $model->{$model->getKeyName()} = Media::forPathOnDisk(
+                        $model->disk, $model->getDiskPath())
+                        ->pluck($model->getKeyName())->first();
+                    $model->exists = true;
+                    break;
+                case static::ON_DUPLICATE_INCREMENT:
+                default:
+                    $model->filename = $this->generateUniqueFilename($model);
+            }
         }
+        return $model;
     }
 
     /**
@@ -694,6 +701,16 @@ class MediaUploader
             ->where('extension', $model->extension)
             ->delete();
 
+        $this->deleteExistingFile($model);
+    }
+
+    /**
+     * Delete the file on disk.
+     * @param  \Plank\Mediable\Media  $model
+     * @return void
+     */
+    private function deleteExistingFile(Media $model)
+    {
         $this->filesystem->disk($model->disk)->delete($model->getDiskPath());
     }
